refactor: merge blocks and inventory slices to share the state
