+++ src/store/slices/blocksSlice.js
@@ -2,28 +2,78 @@ import { createSlice } from "@reduxjs/toolkit";
 import { blockTypes } from "assets/blockTypes";
 import { nanoid } from "nanoid";
 
-const initialState = [];
+const inventoryItems = Object.values(blockTypes).map((item) => ({
+  name: item,
+  amount: 64,
+}));
+
+const initialState = {
+  blocksPositions: [],
+  inventory: { selectedItem: 1, inventoryItems },
+};
+
+export const inventoryActionTypes = {
+  INCREASE_SELECTED_ITEM: "INCREASE_SELECTED_ITEM",
+  DECREASE_SELECTED_ITEM: "DECREASE_SELECTED_ITEM",
+};
+
+export const MAX_INVENTORY_ITEMS = 9;
 
 export const blocksSlice = createSlice({
   name: "blocks",
   initialState,
   reducers: {
     addBlock: (state, action) => {
-      return [
+      return {
         ...state,
-        {
-          id: nanoid(),
-          position: action.payload.position,
-          type: blockTypes.DIRT,
-        },
-      ];
+        blocksPositions: [
+          ...state.blocksPositions,
+          {
+            id: nanoid(),
+            position: action.payload.position,
+            type: blockTypes.DIRT,
+          },
+        ],
+      };
     },
     removeBlock: (state, action) => {
-      return [...state].filter((item) => item.id !== action.payload.id);
+      return {
+        ...state,
+        blocksPositions: [...state.blocksPositions].filter(
+          (item) => item.id !== action.payload.id
+        ),
+      };
+    },
+    changeSelectedItem: (state, action) => {
+      const { type } = action.payload;
+      const { INCREASE_SELECTED_ITEM, DECREASE_SELECTED_ITEM } =
+        inventoryActionTypes;
+      const { selectedItem } = state.inventory;
+      if (type === INCREASE_SELECTED_ITEM) {
+        const newSelectedItem =
+          selectedItem === MAX_INVENTORY_ITEMS - 1 ? 0 : selectedItem + 1;
+        return {
+          ...state,
+          inventory: { ...state.inventory, selectedItem: newSelectedItem },
+        };
+      }
+      if (type === DECREASE_SELECTED_ITEM) {
+        const newSelectedItem =
+          selectedItem === 0 ? MAX_INVENTORY_ITEMS - 1 : selectedItem - 1;
+        return {
+          ...state,
+          inventory: { ...state.inventory, selectedItem: newSelectedItem },
+        };
+      }
+      return state;
+    },
+    changeBlocksAmount: (state) => {
+      return state;
     },
   },
 });
 
-export const { addBlock, removeBlock } = blocksSlice.actions;
+export const { addBlock, removeBlock, changeSelectedItem, changeBlocksAmount } =
+  blocksSlice.actions;
 
 export default blocksSlice.reducer;
